+++ can_library/configs/nodes/DASHBOARD.json
@@ -37,7 +37,7 @@
                             "type": "bool"
                         }
                     ],
-                    "msg_period": 300,
+                    "msg_period": 200,
                     "msg_priority": 1
                 },
                 {
