+++ source/dashboard/CMakeLists.txt
@@ -4,5 +4,5 @@ create_can_node_lib(DASHBOARD "CMSIS_G474;FREERTOS_G474")
 add_firmware_component(
     NAME dashboard
     LINKER_SCRIPT "STM32G474RETX_FLASH"
-    LIBS "can_node_DASHBOARD;CMSIS_G474;FREERTOS_G474;PHAL_G474;STRBUF;HEARTBEAT"
+    LIBS "can_node_DASHBOARD;CMSIS_G474;FREERTOS_G474;PHAL_G474;STRBUF;HEARTBEAT;NEXTION"
 )
